Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #9:

This patch reduces the compiler warning count on Win64 by an additional
350, as reported by VS2013.

+++ sql/item_func.h
@@ -2067,7 +2067,7 @@ public:
   bool fix_fields(THD *thd, Item **ref);
   virtual void print(String *str, enum_query_type query_type);
   void set_null_value(const CHARSET_INFO* cs);
-  void set_value(const char *str, uint length, const CHARSET_INFO* cs);
+  void set_value(const char *str, size_t length, const CHARSET_INFO* cs);
 };
